+++ src/main/java/com/extendedclip/deluxemenus/DeluxeMenus.java
@@ -7,7 +7,7 @@ import com.extendedclip.deluxemenus.dupe.DupeFixer;
 import com.extendedclip.deluxemenus.dupe.MenuItemMarker;
 import com.extendedclip.deluxemenus.hooks.*;
 import com.extendedclip.deluxemenus.listener.PlayerListener;
-import com.extendedclip.deluxemenus.menu.HeadType;
+import com.extendedclip.deluxemenus.menu.options.HeadType;
 import com.extendedclip.deluxemenus.menu.Menu;
 import com.extendedclip.deluxemenus.metrics.Metrics;
 import com.extendedclip.deluxemenus.nbt.NbtProvider;
