+++ src/main/java/com/extendedclip/deluxemenus/menu/options/MenuItemOptions.java
@@ -2,7 +2,6 @@ package com.extendedclip.deluxemenus.menu.options;
 
 import com.extendedclip.deluxemenus.action.ClickHandler;
 import com.extendedclip.deluxemenus.config.DeluxeMenusConfig;
-import com.extendedclip.deluxemenus.menu.HeadType;
 import com.extendedclip.deluxemenus.requirement.RequirementList;
 import org.bukkit.DyeColor;
 import org.bukkit.block.banner.Pattern;
